update attendanceRecording done~

+++ CStoreManagement/app/src/main/java/com/cstore/zhiyazhang/cstoremanagement/bean/WorkHoursBean.kt
@@ -15,5 +15,5 @@ data class WorkHoursBean(
         /**
          * 总小时
          */
-        val hours: Double
+        val hours: Int
 ) : Serializable
